refactor: move insert_globals into SQLUpgradeService static method
